requirement.txt
certifi==2018.8.13
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1
django-redis==4.9.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
